Add full-text search for items endpoint
SQLite version of Node.js 23.5 does not came with the FTS5 extension by default. It is necessary to build then load the shared library at runtime. The file `setup.sh` downloads and builds the extension.

In addition to the `search` service and the respective HTTP endpoint, it was also necessary to modify the database module and to add triggers to maintain consistency between `items` and `items_fts` tables.